Fix twitter card URL
Use absURL function instead of manually concatenating
with Site.BaseURL so that we don't need to manage the optional trailing
slash depending on if we are in a deploy preview (no slash) or not
(slash).

We've been playing whack a mole with this problem and I think this is
the recommended best practice for dealing with the inconsistent slash
going forward.